# api/openapi.yaml
openapi: 3.1.0
info:
  title: SlateDesk API
  description: |
    SlateDesk v2 HTTP API. Milestone 1 surface: local authentication
    (signed HttpOnly session cookie), users, and teams. Milestone 2 adds
    the ticket core: tickets, articles, attachments, tags, dashboard
    counters, and the realtime event stream. Milestone 3 adds the email
    channel's admin surface: mailbox configuration with live fetch/send
    tests and the Google OAuth connect flow, the instance external-URL
    setting, and outbound delivery status on articles with a retry
    operation.

    All errors are RFC 9457 problem details (`application/problem+json`).

    Access model (M2): every ticket-core operation requires an
    authenticated **agent or admin** session — customer access arrives
    with the M4 portal. The one stricter rule: `DELETE /tags/{id}` is
    **admin only**. Agents can additionally **read** users and teams
    (`GET /users`, `/users/{id}`, `/teams`, `/teams/{id}`) — the
    workspace needs them for assignee pickers, team routing, and
    requester lookups; all user/team mutations remain admin only. Each
    operation's description states its required role.

    Access model (M3): everything under `/mailboxes` and `/settings` is
    **admin only** (including the Google OAuth callback, which arrives as
    a top-level browser navigation carrying the admin's session cookie).
    `POST /articles/{id}/retry-send` is agent or admin, like the rest of
...

...
    (`/canned-replies`).

...

...

...
servers:
...
security:
  - cookieAuth: []
  - apiKey: []
...
paths:
...
  /tickets:
    get:
      operationId: listTickets
      summary: List tickets
      description: |
        Filtered, paginated ticket list, most recent activity first
        (`updated_at` descending), or by relevance when `q` is present.
        Requires agent or admin role.

        `view` applies one of the fixed workspace views; every other
        filter composes with it (logical AND):

          - `my` — assigned to the caller, not closed
          - `unassigned` — no assignee, not closed
          - `open` — status is open, waiting_on_customer, or on_hold
          - `closed` — status is closed
      tags: [tickets]
      parameters:
        - name: view
          in: query
          description: Fixed workspace view applied before the other filters.
          schema:
            $ref: "#/components/schemas/TicketView"
        - name: status
          in: query
          description: Statuses to include (repeat the parameter). Empty = all.
          schema:
            type: array
            items:
              $ref: "#/components/schemas/TicketStatus"
        - name: priority
          in: query
          description: Priorities to include (repeat the parameter). Empty = all.
          schema:
            type: array
            items:
              $ref: "#/components/schemas/TicketPriority"
        - name: assignee_id
          in: query
          description: Only tickets assigned to this user.
          schema:
            type: string
            format: uuid
        - name: team_id
          in: query
          description: Only tickets owned by this team.
          schema:
            type: string
            format: uuid
        - name: tag_id
          in: query
          description: Only tickets bearing this tag.
          schema:
            type: string
            format: uuid
        - name: closed_today
          in: query
          description: |
            When true, only tickets closed since midnight (database
            timezone) — the same predicate as the dashboard's
            `closed_today` counter, so that tile can deep-link a queue
            whose contents match its number.
          schema:
            type: boolean
        - name: q
          in: query
          description: |
            Full-text search (websearch syntax, e.g. `printer -toner` or
            `"exact phrase"`) across ticket subjects and article bodies.
            When present, results are ranked by relevance.
          schema:
            type: string
        - name: limit
          in: query
          description: Page size.
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          description: Page offset.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: One page of matching tickets plus the total match count.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createTicket
      summary: Create a ticket
      description: |
        Creates a ticket and its first article in one transaction: the
        day's ticket number (YYYYMMDD-NNNN) is allocated atomically, the
        body is recorded as an agent-authored public article
        (`sender_type=agent`, `channel=web`), and a `created` audit event
        plus a `ticket.created` stream event are emitted. The ticket
        starts `open`; priority defaults to `medium`. Requires agent or
        admin role.
      tags: [tickets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Ticket created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Problem"
...
  /public/tickets:
    post:
      operationId: submitPublicTicket
      summary: Submit a ticket from the public web form
      description: |
        Unauthenticated ticket intake for the single built-in web form
        (channel `web`). The requester is auto-created from (or matched
        to) `email`; the body becomes the first public article. Returns
        **202** with the assigned ticket number.

        Abuse controls: the request is **rate limited per client IP** (429
        when exceeded). `_honeypot` is a hidden field that real users
        leave empty — when it is non-empty the submission is silently
        accepted (still **202**) but no ticket is created, so bots get no
        signal, and `ticket_number` is then absent. An optional captcha
        verification stub is **off by default**; when an operator enables
        it, a missing or invalid token is a 400.
      tags: [public]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicTicketRequest"
      responses:
        "202":
          description: |
            Submission accepted. `ticket_number` is present when a ticket
            was created and absent when the submission was silently
            dropped by an anti-spam control.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTicketAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          description: Too many submissions from this client; retry later.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sd_session
      description: |
        Signed HttpOnly session cookie (SameSite=Lax) carrying the user id
        and token_version. Obtained via POST /auth/login; cleared via
        POST /auth/logout.
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: sd_live_...
      description: |
        Scoped machine API key presented as
        `Authorization: Bearer sd_live_…`. An **alternative** to
        `cookieAuth` on every `/api/v1` operation: because the top-level
        `security` lists the two schemes as separate requirement objects
        (an OR), either scheme alone authenticates a request and the two
        are never both required. The key is a public `sd_live_` prefix
        plus 256 bits of base62 entropy; only its SHA-256 digest is
        stored, and the request acts as the admin or agent who created
        the key, carrying that key's scopes. Scopes are `read` and
        `write` — a `read` key is refused (403) on mutating operations.
        OpenAPI cannot bind scopes to an HTTP-bearer scheme, so the scope
        gate is enforced in the handler and documented rather than
        expressed as security scopes here.
...
  responses:
    BadRequest:
      description: Malformed request body or parameters.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing, invalid, or expired session.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated but not allowed to perform this action.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Problem:
      description: Any other error, as RFC 9457 problem details.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details. Media type application/problem+json.
      required: [title, status]
      properties:
        type:
          type: string
          format: uri-reference
          description: Problem type URI; "about:blank" when unspecified.
          default: "about:blank"
        title:
          type: string
          description: Short human-readable summary of the problem type.
        status:
          type: integer
          description: HTTP status code.
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: Occurrence-specific explanation.
        instance:
          type: string
          format: uri-reference
          description: URI reference identifying this occurrence (request path).
...
    TicketStatus:
      type: string
      description: Ticket lifecycle status.
      enum: [open, waiting_on_customer, on_hold, closed]
    TicketPriority:
      type: string
      description: Ticket priority.
      enum: [low, medium, high, critical]
    TicketView:
      type: string
      description: |
        Fixed workspace views: `my` (assigned to the caller, not closed),
        `unassigned` (no assignee, not closed), `open` (any non-closed
        status), `closed` (closed only).
      enum: [my, unassigned, open, closed]
    ArticleSenderType:
      type: string
      description: Who authored an article.
      enum: [customer, agent, system]
    ArticleChannel:
      type: string
      description: The channel an article arrived through.
      enum: [email, api, web]
    UserSummary:
      type: object
      description: Compact user reference embedded in tickets and articles.
      required: [id, name, email]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
    Ticket:
      type: object
      description: A ticket, as returned by list and update operations.
      required:
        - id
        - number
        - subject
        - status
        - priority
        - requester
        - assignee
        - team_id
        - created_at
        - updated_at
        - closed_at
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
          description: Human-facing ticket number (YYYYMMDD-NNNN).
          pattern: "^[0-9]{8}-[0-9]{4,}$"
        subject:
          type: string
        status:
          $ref: "#/components/schemas/TicketStatus"
        priority:
          $ref: "#/components/schemas/TicketPriority"
        requester:
          $ref: "#/components/schemas/UserSummary"
        assignee:
          description: Assigned agent; null when unassigned.
          anyOf:
            - $ref: "#/components/schemas/UserSummary"
            - type: "null"
        team_id:
          type: ["string", "null"]
          format: uuid
          description: Owning team; null when not routed to a team.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        closed_at:
          type: ["string", "null"]
          format: date-time
          description: Set while the ticket is closed; null otherwise.
    TicketList:
      type: object
      description: One page of tickets plus the pre-pagination match count.
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Ticket"
        total:
          type: integer
          format: int64
          minimum: 0
          description: Total matching tickets before limit/offset.
    TicketDetail:
      description: |
        A ticket with its full thread: tags, every article in
        chronological order (each carrying its attachments), and the
        audit event trail.
      allOf:
        - $ref: "#/components/schemas/Ticket"
        - type: object
          required: [tags, articles, events]
          properties:
            tags:
              type: array
              description: The ticket's tags, sorted by name.
              items:
                $ref: "#/components/schemas/Tag"
            articles:
              type: array
              description: All articles, oldest first.
              items:
                $ref: "#/components/schemas/Article"
            events:
              type: array
              description: Audit trail, oldest first.
              items:
                $ref: "#/components/schemas/TicketEvent"
    Article:
      type: object
      description: |
        One message in a ticket's thread. `body_html`, when present, was
        sanitized (bluemonday) at write time and is safe to render;
        `body_text` always exists and feeds search. `delivery_status` is
        the outbound email badge (null on articles that are not outbound
        email).
      required:
        - id
        - ticket_id
        - author
        - sender_type
        - channel
        - is_internal
        - body_text
        - body_html
        - delivery_status
        - attachments
        - created_at
      properties:
        id:
          type: string
          format: uuid
        ticket_id:
          type: string
          format: uuid
        author:
          description: Authoring user; null for external or system authors.
          anyOf:
            - $ref: "#/components/schemas/UserSummary"
            - type: "null"
        sender_type:
          $ref: "#/components/schemas/ArticleSenderType"
        channel:
          $ref: "#/components/schemas/ArticleChannel"
        is_internal:
          type: boolean
          description: Internal notes are never shown to customers.
        body_text:
          type: string
        body_html:
          type: ["string", "null"]
          description: Sanitized HTML body; null when the article is plain text.
        delivery_status:
          description: |
            Outbound email delivery badge; null on articles that are not
            outbound email (inbound mail, internal notes, web/api
            articles on tickets without an email origin).
          anyOf:
            - $ref: "#/components/schemas/DeliveryStatus"
            - type: "null"
        attachments:
          type: array
          description: Files attached to this article.
          items:
            $ref: "#/components/schemas/Attachment"
        created_at:
          type: string
          format: date-time
    DeliveryStatus:
      type: string
      description: |
        Outbound email delivery state of an article. Transitions:

          - `queued` — the article and its send job committed in one
            transaction; the job has not started yet.
          - `queued → sending` — the send worker committed the outbound
            Message-ID (threading is safe from here on) and is dialing
            SMTP. On worker crash the job retries from here with the
            same Message-ID.
          - `sending → sent` — the SMTP server accepted the message.
            Terminal.
          - `sending → failed` — every retry (capped backoff) failed.
            Terminal, shown as a visible badge; `POST
            /articles/{id}/retry-send` moves it back to `queued`.
          - `sending → queued` — recovery only: a worker crash on the
            final attempt can strand the badge on `sending` (the job is
            discarded without running); retry-send accepts that stuck
            state once no live send job remains.

        Every transition emits an `article.updated` stream event.
      enum: [queued, sending, sent, failed]
    Attachment:
      type: object
      description: |
        Attachment metadata; the bytes live at GET /attachments/{id}.
      required: [id, article_id, filename, content_type, size_bytes, created_at]
      properties:
        id:
          type: string
          format: uuid
        article_id:
          type: string
          format: uuid
        filename:
          type: string
          description: Sanitized original filename.
        content_type:
          type: string
        size_bytes:
          type: integer
          format: int64
          minimum: 0
        created_at:
          type: string
          format: date-time
    TicketEvent:
      type: object
      description: |
        Slim audit event. `payload` is a small type-specific JSON
        document (e.g. `{"from": "open", "to": "closed"}` for
        status_changed).
      required: [id, ticket_id, actor_id, type, payload, created_at]
      properties:
        id:
          type: integer
          format: int64
          description: Insertion-ordered id (stable even within one transaction).
        ticket_id:
          type: string
          format: uuid
        actor_id:
          type: ["string", "null"]
          format: uuid
          description: Acting user; null for system actions.
        type:
          type: string
          description: |
            Event type. One of: created, article_added, status_changed,
            priority_changed, assignee_changed, team_changed, tags_changed.
        payload:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    Tag:
      type: object
      description: A tag. Names are case-insensitively unique.
      required: [id, name, color]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: ["string", "null"]
          description: Display color hint (e.g. "#RRGGBB"); null when unset.
...
    CreateTicketRequest:
      type: object
      required: [subject, body]
      properties:
        subject:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
          description: |
            Plain-text body of the ticket's first article (recorded as an
            agent-authored public web article).
        priority:
          $ref: "#/components/schemas/TicketPriority"
        team_id:
          type: string
          format: uuid
          description: Route the ticket to this team.
        requester_id:
          type: string
          format: uuid
          description: |
            The user the ticket is for (e.g. opening a ticket on a
            customer's behalf). Defaults to the caller when omitted.
...
    PublicTicketRequest:
      type: object
      required: [subject, body, email]
      properties:
        subject:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          description: |
            Submitter's email; the requester is created from or matched to
            it.
        name:
          type: string
          description: Submitter's display name (optional).
        _honeypot:
          type: string
          description: |
            Anti-spam honeypot. Hidden from real users and expected to be
            empty; a non-empty value causes the submission to be silently
            dropped (still 202, no ticket created).
    PublicTicketAccepted:
      type: object
      description: |
        Acknowledgement of a public submission. `ticket_number` is present
        when a ticket was created and absent when an anti-spam control
        silently dropped the submission.
      properties:
        ticket_number:
          type: string
          pattern: "^[0-9]{8}-[0-9]{4,}$"
          description: The assigned ticket number (YYYYMMDD-NNNN).
...